requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
discord-py-slash-command==4.2.1
discord.py==2.4.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
yarl==1.9.4
